fix(tests): silence GCC -Wconversion in cellapp 1000-entity perf test
GCC 13 on Linux CI rejected the implicit conversion from
std::chrono::microseconds::rep (long int) to double at
test_cellapp_integration.cpp:507. Wrap the .count() result in
static_cast<double>(...) before dividing by 1000.0. MSVC was
lenient here, so the warning only showed up on Linux.

+++ tests/integration/test_cellapp_integration.cpp
@@ -504,7 +504,8 @@ TEST(CellAppIntegration, ThousandEntityTickWithinPerfBudget) {
     const auto t0 = std::chrono::steady_clock::now();
     space.Tick(0.1f);
     const auto t1 = std::chrono::steady_clock::now();
-    tick_ms.push_back(std::chrono::duration_cast<std::chrono::microseconds>(t1 - t0).count() /
+    tick_ms.push_back(static_cast<double>(
+                          std::chrono::duration_cast<std::chrono::microseconds>(t1 - t0).count()) /
                       1000.0);
   }
   std::sort(tick_ms.begin(), tick_ms.end());
